requirements.txt
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysql-connector==2.2.9
python-dotenv==0.10.3
six==1.13.0
Werkzeug==0.16.0
